ZON-2898: Move test to where it belongs
