+++ src/zeit/content/article/edit/browser/tests/test_reference.py
@@ -48,6 +48,38 @@ class ImageForm(zeit.content.article.edit.browser.testing.BrowserTestCase):
         b.getControl('Apply').click()
         self.assertFalse(self.get_image_block().set_manually)
 
+    def test_png_teaser_images_should_enable_colorpicker(self):
+        from zeit.content.image.testing import (
+            create_image_group_with_master_image)
+        with zeit.cms.testing.site(self.getRootFolder()):
+            with zeit.cms.testing.interaction():
+                article = zeit.content.article.testing.create_article()
+                group = create_image_group_with_master_image(
+                    file_name='http://xml.zeit.de/2016/DSC00109_2.PNG')
+                zeit.content.image.interfaces.IImages(article).image = group
+                self.repository['article'] = article
+        b = self.browser
+        b.open(
+            'http://localhost/++skin++vivi/repository/article/@@checkout')
+        b.open('@@edit.form.teaser-image?show_form=1')
+        b.getControl(name='teaser-image.fill_color')
+
+    def test_non_png_teaser_images_should_not_enable_colorpicker(self):
+        from zeit.content.image.testing import (
+            create_image_group_with_master_image)
+        with zeit.cms.testing.site(self.getRootFolder()):
+            with zeit.cms.testing.interaction():
+                article = zeit.content.article.testing.create_article()
+                group = create_image_group_with_master_image()
+                zeit.content.image.interfaces.IImages(article).image = group
+                self.repository['article'] = article
+        b = self.browser
+        b.open(
+            'http://localhost/++skin++vivi/repository/article/@@checkout')
+        b.open('@@edit.form.teaser-image?show_form=1')
+        with self.assertRaises(LookupError):
+            b.getControl(name='teaser-image.fill_color')
+
 
 class ImageEditTest(zeit.content.article.edit.browser.testing.EditorTestCase):
